+++ CHANGELOG.md
@@ -4,12 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2019-11-28
+### Added
+- Auto retry when a network is found (only if disconnected before).
+
+### Changed
+- Minimum SDK version is now JellyBeans (Android 4.1, SDK 16),
+- Updated the connection mode to a modern Worker.
+
+### Fixed
+- Unable tu display activity on SDK 19.
+
 ## [1.0.0] - 2019-11-02
 ### Added
-- Base application and widget.
-- IP finder with [ipify API][ipify].
+- Base application and widget,
+- IP finder with [ipify API][ipify],
 - Application icon based on the work of [Eugenio Hansen][globe], licensed under [CC BY-SA 3.0].
 
+[1.0.1]: https://github.com/guildem/publicip-android/releases/tag/1.0.1
 [1.0.0]: https://github.com/guildem/publicip-android/releases/tag/1.0.0
 
 [ipify]: https://www.ipify.org/